chore(validator): migrate footer
