+++ core/shared-ui/src/commonMain/kotlin/com/neoutils/neoregex/core/sharedui/extension/PointerInputScope.kt
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.neoutils.neoregex.feature.matcher.extension
+package com.neoutils.neoregex.core.sharedui.extension
 
 import androidx.compose.foundation.gestures.awaitEachGesture
 import androidx.compose.foundation.gestures.awaitFirstDown
