+++ feature/about/src/commonMain/kotlin/com/neoutils/neoregex/feature/about/screen/AboutScreen.kt
@@ -19,8 +19,10 @@
 package com.neoutils.neoregex.feature.about.screen
 
 import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.MaterialTheme.colorScheme
 import androidx.compose.material3.MaterialTheme.typography
 import androidx.compose.material3.ProvideTextStyle
 import androidx.compose.material3.Text
@@ -51,6 +53,7 @@ class AboutScreen : Screen {
     @Composable
     override fun Content() = Column(
         modifier = Modifier
+            .background(colorScheme.background)
             .padding(dimensions.default)
             .fillMaxSize(),
         verticalArrangement = Arrangement.spacedBy(
