+++ feature/matcher/src/commonMain/kotlin/com/neoutils/neoregex/feature/matcher/MatcherScreen.kt
@@ -19,51 +19,29 @@
 package com.neoutils.neoregex.feature.matcher
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CornerBasedShape
-import androidx.compose.foundation.shape.CornerSize
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Icon
-import androidx.compose.material3.LocalContentColor
+import androidx.compose.foundation.layout.BoxWithConstraints
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material3.MaterialTheme.colorScheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.VerticalDivider
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.focus.focusProperties
-import androidx.compose.ui.focus.onFocusChanged
-import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.input.key.onPreviewKeyEvent
 import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import cafe.adriel.voyager.core.model.rememberScreenModel
 import cafe.adriel.voyager.core.screen.Screen
-import com.neoutils.highlight.compose.remember.rememberTextFieldValue
+import com.neoutils.neoregex.core.common.extension.toTextState
+import com.neoutils.neoregex.core.common.model.Target
 import com.neoutils.neoregex.core.common.util.Command
-import com.neoutils.neoregex.core.designsystem.textfield.NeoTextField
-import com.neoutils.neoregex.core.designsystem.theme.NeoTheme.dimensions
-import com.neoutils.neoregex.core.resources.Res
-import com.neoutils.neoregex.core.resources.ic_redo_24
-import com.neoutils.neoregex.core.resources.ic_undo_24
-import com.neoutils.neoregex.core.resources.matcher_footer_insert_regex_hint
+import com.neoutils.neoregex.core.sharedui.component.Footer
 import com.neoutils.neoregex.core.sharedui.component.Performance
 import com.neoutils.neoregex.core.sharedui.component.TextEditor
 import com.neoutils.neoregex.feature.matcher.action.MatcherAction
-import com.neoutils.neoregex.feature.matcher.extension.onLongHold
-import com.neoutils.neoregex.feature.matcher.extension.toTextState
-import com.neoutils.neoregex.feature.matcher.model.Syntax
-import com.neoutils.neoregex.feature.matcher.model.Target
 import com.neoutils.neoregex.feature.matcher.state.MatcherUiState
-import org.jetbrains.compose.resources.painterResource
-import org.jetbrains.compose.resources.stringResource
 
 class MatcherScreen : Screen {
 
@@ -84,7 +62,7 @@ class MatcherScreen : Screen {
                 value = uiState.inputs.text,
                 onValueChange = {
                     viewModel.onAction(
-                        MatcherAction.Input.UpdateText(it.toTextState())
+                        MatcherAction.UpdateText(it.toTextState())
                     )
                 },
                 onFocusChange = {
@@ -129,175 +107,23 @@ class MatcherScreen : Screen {
         }
 
         Footer(
-            uiState = uiState,
+            inputs = uiState.inputs,
+            history = uiState.history,
             onAction = viewModel::onAction,
+            error = remember(uiState.result) {
+                when (val result = uiState.result) {
+                    is MatcherUiState.Result.Failure -> result.error
+                    is MatcherUiState.Result.Success -> ""
+                }
+            },
+            onFocus = {
+                if (it.isFocused) {
+                    viewModel.onAction(
+                        MatcherAction.TargetChange(Target.REGEX)
+                    )
+                }
+            },
             modifier = Modifier.fillMaxWidth(),
-        )
-    }
-
-    @Composable
-    private fun Footer(
-        uiState: MatcherUiState,
-        onAction: (MatcherAction) -> Unit,
-        syntax: Syntax.Regex = Syntax.Regex(),
-        modifier: Modifier = Modifier
-    ) = Surface(
-        modifier = modifier,
-        shape = RectangleShape,
-        shadowElevation = dimensions.small,
-        color = colorScheme.surfaceContainer,
-        contentColor = colorScheme.onSurface,
-    ) {
-
-        Row(Modifier.fillMaxWidth()) {
-            NeoTextField(
-                value = syntax
-                    .highlight
-                    .rememberTextFieldValue(uiState.inputs.regex),
-                onValueChange = {
-                    onAction(
-                        MatcherAction.Input.UpdateRegex(
-                            it.toTextState(allowMultiline = false)
-                        )
-                    )
-                },
-                singleLine = true,
-                modifier = Modifier
-                    .align(Alignment.CenterVertically)
-                    .weight(weight = 1f)
-                    .onFocusChanged {
-                        if (it.isFocused) {
-                            onAction(
-                                MatcherAction.TargetChange(Target.REGEX)
-                            )
-                        }
-                    }
-                    .onPreviewKeyEvent {
-                        when (Command.from(it)) {
-                            Command.UNDO -> {
-                                onAction(
-                                    MatcherAction.History.Undo(Target.REGEX)
-                                )
-                                true
-                            }
-
-                            Command.REDO -> {
-                                onAction(
-                                    MatcherAction.History.Redo(Target.REGEX)
-                                )
-                                true
-                            }
-
-                            else -> false
-                        }
-                    },
-                hint = stringResource(Res.string.matcher_footer_insert_regex_hint),
-                error = remember(uiState.result) {
-                    when (uiState.result) {
-                        is MatcherUiState.Result.Failure -> uiState.result.error
-                        is MatcherUiState.Result.Success -> ""
-                    }
-                }
-            )
-
-            HistoryControl(
-                state = uiState.history,
-                onAction = onAction,
-                modifier = Modifier
-                    .align(Alignment.CenterVertically)
-                    .padding(vertical = dimensions.small)
-                    .padding(end = dimensions.small)
-                    .focusProperties {
-                        canFocus = false
-                    }
-            )
-        }
-    }
-
-    @Composable
-    private fun HistoryControl(
-        state: MatcherUiState.History,
-        onAction: (MatcherAction.History) -> Unit,
-        modifier: Modifier = Modifier,
-        shape: CornerBasedShape = RoundedCornerShape(dimensions.small)
-    ) = Row(
-        modifier = modifier
-            .height(IntrinsicSize.Min)
-            .border(
-                width = 1.dp,
-                color = colorScheme.outline,
-                shape = RoundedCornerShape(dimensions.small)
-            )
-    ) {
-
-        val contentColor = LocalContentColor.current
-
-        Icon(
-            painter = painterResource(Res.drawable.ic_undo_24),
-            contentDescription = null,
-            tint = contentColor.copy(
-                alpha = if (state.canUndo) 1f else 0.5f
-            ),
-            modifier = Modifier
-                .clip(
-                    shape.copy(
-                        topEnd = CornerSize(0.dp),
-                        bottomEnd = CornerSize(0.dp)
-                    )
-                )
-                .clickable(state.canUndo) {
-                    onAction(
-                        MatcherAction.History.Undo()
-                    )
-                }
-                .onLongHold {
-                    onAction(
-                        MatcherAction.History.Undo()
-                    )
-                }
-                .padding(
-                    vertical = dimensions.tiny,
-                    horizontal = dimensions.small,
-                )
-        )
-
-        VerticalDivider(
-            modifier = Modifier
-                .fillMaxHeight()
-                .width(1.dp)
-                .weight(
-                    weight = 1f,
-                    fill = false
-                )
-        )
-
-        Icon(
-            painter = painterResource(Res.drawable.ic_redo_24),
-            contentDescription = null,
-            tint = contentColor.copy(
-                alpha = if (state.canRedo) 1f else 0.5f
-            ),
-            modifier = Modifier
-                .clip(
-                    shape.copy(
-                        topStart = CornerSize(0.dp),
-                        bottomStart = CornerSize(0.dp)
-                    )
-                )
-                .clickable(state.canRedo) {
-                    onAction(
-                        MatcherAction.History.Redo()
-                    )
-                }
-                .onLongHold {
-                    onAction(
-                        MatcherAction.History.Redo()
-                    )
-                }
-                .padding(
-                    vertical = dimensions.tiny,
-                    horizontal = dimensions.small,
-                )
         )
     }
 }
