+++ feature/matcher/src/commonMain/kotlin/com/neoutils/neoregex/feature/matcher/extension/TextState.kt
@@ -21,7 +21,7 @@ package com.neoutils.neoregex.feature.matcher.extension
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.input.TextFieldValue
-import com.neoutils.neoregex.feature.matcher.model.TextState
+import com.neoutils.neoregex.core.common.model.TextState
 
 fun TextState.toTextFieldValue(
     spanStyles: List<AnnotatedString.Range<SpanStyle>> = listOf()
