+++ feature/validator/src/commonMain/kotlin/com/neoutils/neorefex/feature/validator/ValidatorScreen.kt
@@ -18,21 +18,40 @@
 
 package com.neoutils.neorefex.feature.validator
 
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme.colorScheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import cafe.adriel.voyager.core.screen.Screen
+import com.neoutils.neoregex.core.common.model.Inputs
+import com.neoutils.neoregex.core.sharedui.component.Footer
+import com.neoutils.neoregex.core.sharedui.model.History
 
 class ValidatorScreen : Screen {
 
     @Composable
-    override fun Content() = Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
+    override fun Content() = Column(
+        modifier = Modifier
+            .background(colorScheme.background)
+            .fillMaxSize()
     ) {
-        Text(text = "Hello, World!")
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .weight(1f),
+            contentAlignment = Alignment.Center
+        ) {
+            Text(text = "Hello, World!")
+        }
+
+        Footer(
+            inputs = Inputs(),
+            history = History(),
+        )
     }
 }
